+++ luadatapacktypes/src/entity.d.ts
@@ -117,4 +117,21 @@ interface Entity {
      * @param obj Nbt object
      */
     merge_nbt(obj : NbtCompound) : void;
+
+    /**
+     * Applies an effect to the entity
+     * 
+     * @param effect Effect
+     * @param duration Duration in seconds
+     * @param amplifier Amplifier
+     * @param hide_particles Hides effect particles
+     */
+    add_effect(effect : string, duration : number, amplifier : number, hide_particles : boolean) : void;
+
+    /**
+     * Clears all effects
+     * 
+     * @remarks It calls `/effect clear`
+     */
+    clear_effects() : void;
 }